feat: Ethereum indexer class
