+++ backend/api/.env.example
@@ -7,3 +7,5 @@ GOERLI_URL=
 SOLANA_ADDRESS=
 ETH_ADDRESS=
 INFURA_PROJECT_ID=
+SEPOLIA_URL=
+SEPOLIA_URL_HTTP=
