+++ backend/api/src/cluster.ts
@@ -1,0 +1,17 @@
+import { PORT } from "@paybox/common";
+import cluster from "cluster";
+import os from "os";
+import { server } from ".";
+
+const cpuCount = os.cpus().length;
+
+if (cluster.isMaster) {   // isPrimary
+  // Create a worker for each CPU
+  for (let i = 0; i < cpuCount; i++) {
+    cluster.fork();
+  }
+} else {
+  server.listen(PORT, () => {
+    console.log(`Server listening on port: ${PORT}`);
+  });
+}
