+++ backend/api/src/types/chain.ts
@@ -1,0 +1,10 @@
+export enum EthNetwok {
+    sepolia = "sepolia",
+    mainnet = "mainnet",
+    goerli = "goerli"
+}
+
+export type Filter = {
+    address: string,
+    topics: []
+}
